requirements.txt
alembic==1.8.1
anyio==3.6.1
bcrypt==4.0.0
black==22.8.0
click==8.1.3
fastapi==0.82.0
FastAPI-SQLAlchemy==0.2.1
flake8==5.0.4
h11==0.13.0
idna==3.3
isort==5.10.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
